+++ lemoncheesecake/matching/matchers/types_.py
@@ -21,6 +21,8 @@ _TYPE_NAMES = {
 if six.PY2:
     _TYPE_NAMES[unicode] = "string",
 
+STRING_TYPES = [str] if six.PY3 else [str, unicode]
+
 
 class IsValueOfType(Matcher):
     def __init__(self, types, type_name, value_matcher):
@@ -65,6 +67,6 @@ def _is_type(types, type_name):
 is_integer = _is_type([int], "an integer")
 is_float = _is_type([float], "a float")
 is_bool = _is_type([bool], "a boolean")
-is_str = _is_type([str] if six.PY3 else [str, unicode], "a string")
+is_str = _is_type(STRING_TYPES, "a string")
 is_dict = _is_type([dict], "a collection")
 is_list = _is_type([list, tuple], "a list")